+++ mosaic-tty/src/commonMain/c/mosaic-test-tty.h
@@ -13,15 +13,15 @@ typedef struct MosaicTestTtyInitResult {
 	bool already_bound;
 } MosaicTestTtyInitResult;
 
-MosaicTestTtyInitResult testTty_init();
-MosaicTty *testTty_getTty(MosaicTestTty *testTty);
-MosaicTtyIoResult testTty_write(MosaicTestTty *testTty, uint8_t *buffer, int count);
-MosaicTtyIoResult testTty_read(MosaicTestTty *testTty, uint8_t *buffer, int count);
-uint32_t testTty_interruptRead(MosaicTestTty *testTty);
-uint32_t testTty_resize(MosaicTestTty *testTty, int columns, int rows, int width, int height);
-uint32_t testTty_sendFocusEvent(MosaicTestTty *testTty, bool focused);
-uint32_t testTty_sendKeyEvent(MosaicTestTty *testTty);
-uint32_t testTty_sendMouseEvent(MosaicTestTty *testTty);
-uint32_t testTty_free(MosaicTestTty *testTty);
+MOSAIC_EXPORT MosaicTestTtyInitResult testTty_init();
+MOSAIC_EXPORT MosaicTty *testTty_getTty(MosaicTestTty *testTty);
+MOSAIC_EXPORT MosaicTtyIoResult testTty_write(MosaicTestTty *testTty, uint8_t *buffer, int count);
+MOSAIC_EXPORT MosaicTtyIoResult testTty_read(MosaicTestTty *testTty, uint8_t *buffer, int count);
+MOSAIC_EXPORT uint32_t testTty_interruptRead(MosaicTestTty *testTty);
+MOSAIC_EXPORT uint32_t testTty_resize(MosaicTestTty *testTty, int columns, int rows, int width, int height);
+MOSAIC_EXPORT uint32_t testTty_sendFocusEvent(MosaicTestTty *testTty, bool focused);
+MOSAIC_EXPORT uint32_t testTty_sendKeyEvent(MosaicTestTty *testTty);
+MOSAIC_EXPORT uint32_t testTty_sendMouseEvent(MosaicTestTty *testTty);
+MOSAIC_EXPORT uint32_t testTty_free(MosaicTestTty *testTty);
 
 #endif // MOSAIC_TEST_TTY_H